Change how header rows are calculated
Analyze each cell in a row to determine how many numbers are in that
cell and require there to be at least 80% of the cells to be "numeric
deficient". That is currently defined as being a minimum of 75% of the
characters being non-numeric. Cursory checking shows this algorithm
works better that the old one.